+++ shared/viewmodels/CalendarViewModel.js
